Extract WebSocket connection to websocket-connection.js
Move WebSocket connection lifecycle, message handlers, and effects
to a dedicated module using dependency injection pattern.

Features extracted:
- Pure WebSocket message handlers (functional core)
- Effect executor (side effects at edges)
- Connection management with exponential backoff
- Visibility change reconnection logic

Also fixes loadTokens() bug - function was called but never imported.

Reduces app.js by ~215 lines (790 → 575, 27% reduction).
