+++ public/app.js
@@ -14,7 +14,7 @@
     import { createSessionStore, invalidateSessions } from "/lib/session-store.js";
     import { createSessionListComponent } from "/lib/session-list-component.js";
     import { createSessionManager } from "/lib/session-manager.js";
-    import { createTokenStore, setNewToken, invalidateTokens } from "/lib/token-store.js";
+    import { createTokenStore, setNewToken, invalidateTokens, loadTokens as reloadTokens } from "/lib/token-store.js";
     import { createTokenListComponent } from "/lib/token-list-component.js";
     import { createTokenFormManager } from "/lib/token-form.js";
     import { createShortcutsStore } from "/lib/shortcuts-store.js";
@@ -40,6 +40,7 @@
     import { loadQRLib, getConnectInfo, checkPairingStatus } from "/lib/wizard-utils.js";
     import { initModals } from "/lib/modal-init.js";
     import { createViewportManager } from "/lib/viewport-manager.js";
+    import { createWebSocketConnection } from "/lib/websocket-connection.js";
 
     // --- Modal Manager ---
     const modals = new ModalRegistry();
@@ -209,241 +210,7 @@
     });
     terminalKeyboard.init();
 
-    // --- Pure WebSocket message handlers (functional core) ---
-    const wsMessageHandlers = {
-      attached: (msg, currentState) => ({
-        stateUpdates: {
-          'connection.attached': true,
-          'scroll.userScrolledUpBeforeDisconnect': false
-        },
-        effects: [
-          { type: 'updateP2PIndicator' },
-          { type: 'initP2P' },
-          { type: 'scrollToBottomIfNeeded', condition: !currentState.scroll.userScrolledUpBeforeDisconnect }
-        ]
-      }),
-
-      output: (msg) => ({
-        stateUpdates: {},
-        effects: [
-          { type: 'terminalWrite', data: msg.data, preserveScroll: true }
-        ]
-      }),
-
-      'p2p-signal': (msg, currentState) => ({
-        stateUpdates: {},
-        effects: currentState.p2p.peer
-          ? [{ type: 'p2pSignal', data: msg.data }]
-          : []
-      }),
-
-      'p2p-ready': () => ({
-        stateUpdates: {},
-        effects: [
-          { type: 'log', message: '[P2P] Server confirmed DataChannel ready' },
-          { type: 'updateP2PIndicator' }
-        ]
-      }),
-
-      'p2p-closed': () => ({
-        stateUpdates: { 'p2p.connected': false },
-        effects: [
-          { type: 'log', message: '[P2P] Server reports DataChannel closed' },
-          { type: 'updateP2PIndicator' }
-        ]
-      }),
-
-      'pair-complete': (msg, currentState, wizardActivePairCode) => ({
-        stateUpdates: {},
-        effects: (wizardActivePairCode && msg.code === wizardActivePairCode)
-          ? [{ type: 'stopWizardPairing' }, { type: 'switchSettingsView', view: 'success' }]
-          : []
-      }),
-
-      reload: () => ({
-        stateUpdates: {},
-        effects: [{ type: 'reload' }]
-      }),
-
-      exit: () => ({
-        stateUpdates: {},
-        effects: [{ type: 'terminalWrite', data: '\r\n[shell exited]\r\n' }]
-      }),
-
-      'session-removed': () => ({
-        stateUpdates: {},
-        effects: [{ type: 'terminalWrite', data: '\r\n[session deleted]\r\n' }]
-      }),
-
-      'session-renamed': (msg) => ({
-        stateUpdates: { 'session.name': msg.name },
-        effects: [{ type: 'updateSessionUI', name: msg.name }]
-      }),
-
-      'credential-registered': () => ({
-        stateUpdates: {},
-        effects: [{ type: 'refreshTokensAfterRegistration' }]
-      })
-    };
-
-    // Effect executor (side effects at edges)
-    function executeEffect(effect) {
-      switch (effect.type) {
-        case 'updateP2PIndicator':
-          updateP2PIndicator();
-          break;
-        case 'initP2P':
-          p2pManager.create();
-          break;
-        case 'scrollToBottomIfNeeded':
-          if (effect.condition) {
-            scrollToBottom(term);
-          }
-          break;
-        case 'terminalWrite':
-          if (effect.preserveScroll) {
-            terminalWriteWithScroll(term, effect.data);
-          } else {
-            term.write(effect.data);
-          }
-          break;
-        case 'p2pSignal':
-          p2pManager.signal(effect.data);
-          break;
-        case 'log':
-          console.log(effect.message);
-          break;
-        case 'stopWizardPairing':
-          stopWizardPairing();
-          break;
-        case 'switchSettingsView':
-          switchSettingsView(effect.view === 'success' ? viewSuccess : null);
-          break;
-        case 'reload':
-          location.reload();
-          break;
-        case 'updateSessionUI':
-          document.title = effect.name;
-          const url = new URL(window.location);
-          url.searchParams.set("s", effect.name);
-          history.replaceState(null, "", url);
-          renderBar(effect.name);
-          break;
-        case 'refreshTokensAfterRegistration':
-          // Refresh token list to show newly used token
-          loadTokens();
-          // Hide token creation form and show "Generate New Token" button
-          const tokenCreateForm = document.getElementById("token-create-form");
-          const createTokenBtn = document.getElementById("settings-create-token");
-          if (tokenCreateForm) tokenCreateForm.style.display = "none";
-          if (createTokenBtn) createTokenBtn.style.display = "block";
-          break;
-      }
-    }
-
-    let isConnecting = false;
-    let reconnectTimeout = null;
-
-    function connect() {
-      // Prevent multiple simultaneous connection attempts
-      if (isConnecting) {
-        console.log('[WS] Already connecting, skipping duplicate attempt');
-        return;
-      }
-
-      // Clear any pending reconnection timeout
-      if (reconnectTimeout) {
-        clearTimeout(reconnectTimeout);
-        reconnectTimeout = null;
-      }
-
-      isConnecting = true;
-      const proto = location.protocol === "https:" ? "wss:" : "ws:";
-      state.connection.ws = new WebSocket(`${proto}//${location.host}`);
-
-      state.connection.ws.onopen = () => {
-        isConnecting = false;
-        state.connection.reconnectDelay = 1000;
-        state.connection.ws.send(JSON.stringify({ type: "attach", session: state.session.name, cols: term.cols, rows: term.rows }));
-        state.connection.ws.send(JSON.stringify({ type: "resize", cols: term.cols, rows: term.rows }));
-      };
-
-      state.connection.ws.onmessage = (e) => {
-        const msg = JSON.parse(e.data);
-        const handler = wsMessageHandlers[msg.type];
-
-        if (handler) {
-          const { stateUpdates, effects } = handler(msg, state, wizardActivePairCode);
-
-          // Apply state updates
-          if (Object.keys(stateUpdates).length > 0) {
-            state.updateMany(stateUpdates);
-          }
-
-          // Execute effects
-          effects.forEach(executeEffect);
-        }
-      };
-
-      state.connection.ws.onclose = (event) => {
-        isConnecting = false;
-
-        // Check if connection was closed due to revoked credentials
-        if (event.code === 1008) { // 1008 = Policy Violation
-          console.log('[Auth] Session invalidated, redirecting to login');
-          // Redirect to login with message
-          window.location.href = '/login?reason=revoked';
-          return;
-        }
-
-        // Normal disconnect - attempt reconnection with exponential backoff
-        const viewport = document.querySelector(".xterm-viewport");
-        state.scroll.userScrolledUpBeforeDisconnect = !isAtBottom(viewport);
-        state.connection.attached = false;
-        p2pManager.destroy();
-
-        console.log(`[WS] Reconnecting in ${state.connection.reconnectDelay}ms`);
-        reconnectTimeout = setTimeout(connect, state.connection.reconnectDelay);
-        state.connection.reconnectDelay = Math.min(state.connection.reconnectDelay * 2, 10000);
-      };
-
-      state.connection.ws.onerror = (err) => {
-        console.log('[WS] Connection error:', err.message || 'Unknown error');
-        isConnecting = false;
-        state.connection.ws.close();
-      };
-    }
-
-    // Force reconnect when returning to PWA after being backgrounded
-    let hiddenAt = 0;
-    document.addEventListener("visibilitychange", () => {
-      if (document.hidden) {
-        hiddenAt = Date.now();
-      } else {
-        // Coming back to foreground
-        const hiddenDuration = Date.now() - hiddenAt;
-
-        // Skip if already connecting
-        if (isConnecting) {
-          console.log('[Reconnect] Already connecting, skipping visibility reconnect');
-          return;
-        }
-
-        // If was hidden for more than 5 seconds, force reconnect
-        if (hiddenDuration > 5000 && state.connection.ws && !isConnecting) {
-          console.log(`[Reconnect] Was hidden for ${Math.round(hiddenDuration/1000)}s, forcing reconnect`);
-          state.connection.ws.close();
-        } else if (state.connection.ws && state.connection.ws.readyState === WebSocket.OPEN) {
-          // Quick test - send resize to verify connection is alive
-          try {
-            state.connection.ws.send(JSON.stringify({ type: "resize", cols: term.cols, rows: term.rows }));
-          } catch (e) {
-            console.log("[Reconnect] Send failed, forcing reconnect");
-            state.connection.ws.close();
-          }
-        }
-      }
-    });
+    // WebSocket connection setup moved to after all dependencies are initialized (see before Boot section)
 
     // --- Layout ---
 
@@ -640,6 +407,7 @@
     // --- Token management ---
 
     const tokenStore = createTokenStore();
+    const loadTokens = () => reloadTokens(tokenStore);
 
     // Create token form manager with callbacks
     const tokenFormManager = createTokenFormManager({
@@ -778,10 +546,27 @@
     });
     networkMonitor.init();
 
+    // --- WebSocket Connection ---
+
+    const wsConnection = createWebSocketConnection({
+      term,
+      state,
+      p2pManager,
+      updateP2PIndicator,
+      stopWizardPairing,
+      switchSettingsView,
+      viewSuccess,
+      loadTokens,
+      getWizardActivePairCode: () => window.wizardActivePairCode,
+      isAtBottom,
+      renderBar
+    });
+    wsConnection.initVisibilityReconnect();
+
     // --- Boot ---
 
     renderBar(state.session.name);  // Initial render
-    connect();
+    wsConnection.connect();
     loadShortcuts();
     term.focus();
